pyproject.toml
[project]
name = "ai-slogan-maker"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "openai>=2.24.0",
    "python-dotenv>=1.2.2",
]
